testing/web-platform/tests/webrtc-extensions/RTCRtpReceiver-playoutDelay.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Tests for RTCRtpReceiver-playoutDelay attribute</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
'use strict'

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
}, 'audio playoutDelay is null by default');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 0.5;
  assert_equals(receiver.playoutDelay, 0.5);
}, 'audio playoutDelay accepts posititve values');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 4.0;
  assert_throws_js(RangeError, () => {
    receiver.playoutDelay = 4.001;
  }, 'audio playoutDelay doesn\'t accept values greater than 4 seconds');
  assert_equals(receiver.playoutDelay, 4.0);
}, 'audio playoutDelay accepts values up to 4 seconds');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 0.7
  assert_throws_js(RangeError, () => {
    receiver.playoutDelay = -0.5;
  }, 'audio playoutDelay doesn\'t accept negative values');
  assert_equals(receiver.playoutDelay, 0.7);
}, 'audio playoutDelay returns last valid value on throw');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 0.0;
  assert_equals(receiver.playoutDelay, 0.0);
}, 'audio playoutDelay allows zero value');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 0.5;
  assert_equals(receiver.playoutDelay, 0.5);
  receiver.playoutDelay = null;
  assert_equals(receiver.playoutDelay, null);
}, 'audio playoutDelay allows to reset value to null');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
}, 'video playoutDelay is null by default');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 0.5;
  assert_equals(receiver.playoutDelay, 0.5);
}, 'video playoutDelay accepts posititve values');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 4.0;
  assert_throws_js(RangeError, () => {
    receiver.playoutDelay = 4.001;
  }, 'video playoutDelay doesn\'t accept values greater than 4 seconds');
  assert_equals(receiver.playoutDelay, 4.0);
}, 'video playoutDelay accepts values up to 4 seconds');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 0.7
  assert_throws_js(RangeError, () => {
    receiver.playoutDelay = -0.5;
  }, 'video playoutDelay doesn\'t accept negative values');
  assert_equals(receiver.playoutDelay, 0.7);
}, 'video playoutDelay returns last valid value');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 0.0;
  assert_equals(receiver.playoutDelay, 0.0);
}, 'video playoutDelay allows zero value');

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  assert_equals(receiver.playoutDelay, null);
  receiver.playoutDelay = 0.5;
  assert_equals(receiver.playoutDelay, 0.5);
  receiver.playoutDelay = null;
  assert_equals(receiver.playoutDelay, null);
}, 'video playoutDelay allows to reset value to null');
</script>
</body>
